+++ lib/src/date.dart
@@ -8,6 +8,18 @@ import 'date_formatter.dart';
 
 /// Super interface of Jalali and Georgian classes
 abstract class Date {
+  /// Minimum computable julian day number
+  ///
+  /// equivalent to Gregorian(560,3,20) and Jalali(-61,1,1)
+  /// and julian day number of 1925675
+  static const int MIN_JULIAN_DAY_NUMBER = 1925675;
+
+  /// Maximum computable julian day number
+  ///
+  /// equivalent to Gregorian(3798,12,31) and Jalali(3177,10,11)
+  /// and julian day number of 3108616
+  static const int MAX_JULIAN_DAY_NUMBER = 3108616;
+
   /// year
   ///
   /// non-null
@@ -47,9 +59,4 @@ abstract class Date {
   ///
   /// non-null
   bool isLeapYear();
-
-  /// checks if is valid
-  ///
-  /// non-null
-  bool isValid();
 }